Bug Fixes and Improvements
Fix the measure master item preview not displaying on Qlik Cloud
Fix the configuration to manage the custom number format for measures through the config file. Note that the existing measure number format will stay if the custom format is not configured in the config file.
Improve code readability and fix some script logic
Added more samples to the config file